refactor: StrategyServiceをAIディレクトリに移動とimport文更新
- StrategyService.tsとStrategyService.test.tsをapp/src/application/services/ai/に移動
- AIサービス関連ファイルをaiディレクトリに集約し組織化を改善
- 全関連ファイルのimport文パスを新しい場所に合わせて更新
- 相対パス調整とDIコンテナ参照の修正を実施
- リント、ビルド、テスト全てが成功することを確認
