+++ app/src/App.tsx
@@ -4,7 +4,7 @@ import type { AIPort } from './application/ports/AIPort.ts'
 import type { GamePort } from './application/ports/GamePort'
 import type { InputPort } from './application/ports/InputPort'
 import type { PerformanceAnalysisService } from './application/services/PerformanceAnalysisService'
-import type { StrategyService } from './application/services/StrategyService'
+import type { StrategyService } from './application/services/ai/StrategyService'
 import type { GameViewModel } from './application/viewmodels/GameViewModel'
 import type { AIMove, AISettings } from './domain/models/ai/types.ts'
 import { defaultContainer } from './infrastructure/di/DefaultContainer'
