+++ app/src/application/services/ai/StrategyService.ts
@@ -7,12 +7,12 @@ import {
   createStrategyConfig,
   isValidStrategyConfig,
   updateStrategyConfig,
-} from '../../domain/models/ai/StrategyConfig'
+} from '../../../domain/models/ai/StrategyConfig'
 import type {
   StrategyConfig,
   StrategyParameters,
-} from '../../domain/models/ai/StrategyConfig'
-import type { StrategyPort } from '../ports/StrategyPort'
+} from '../../../domain/models/ai/StrategyConfig'
+import type { StrategyPort } from '../../ports/StrategyPort'
 
 /**
  * 戦略統計情報
