+++ app/src/infrastructure/di/DefaultContainer.ts
@@ -9,8 +9,8 @@ import type { TimerPort } from '../../application/ports/TimerPort'
 import { GameApplicationService } from '../../application/services/GameApplicationService'
 import { InputApplicationService } from '../../application/services/InputApplicationService'
 import { PerformanceAnalysisService } from '../../application/services/PerformanceAnalysisService'
-import StrategyService from '../../application/services/StrategyService'
 import { MLAIService } from '../../application/services/ai/MLAIService'
+import StrategyService from '../../application/services/ai/StrategyService'
 import * as ChainDetectionService from '../../domain/services/ChainDetectionService'
 import { PuyoSpawningService } from '../../domain/services/PuyoSpawningService'
 import { BrowserTimerAdapter } from '../adapters/BrowserTimerAdapter'
